improve sample data generation for CD Baby
